DC-11 Move the logger component into decoupled-logger package
