+++ packages/decoupled-logger/package.json
@@ -1,0 +1,36 @@
+{
+  "name": "decoupled-logger",
+  "version": "1.0.1",
+  "description": "decoupled logger.",
+  "main": "index.js",
+  "license": "MIT",
+  "scripts": {
+    "build": "scripts/build.sh",
+    "install": "scripts/build.sh",
+    "watch": "tsc -p tsconfig.dev.json --watch",
+    "lint": "tslint -t verbose src/**/*.ts{,x}",
+    "test": "jest dist"
+  },
+  "engines": {
+    "node": "node >= 8.0.0"
+  },
+  "dependencies": {
+    "@types/jest": "^22.2.3",
+    "@types/lodash": "^4.14.109",
+    "@types/winston": "^2.3.9",
+    "body-parser": "^1.18.2",
+    "chalk": "^2.4.1",
+    "lodash": "4.x",
+    "node-cache": "^4.1.1",
+    "node-fetch": "^1.7.2",
+    "promise.map": "^0.0.2",
+    "tslint": "^5.10.0",
+    "typescript": "^2.8.3",
+    "winston": "^3.0.0"
+  },
+  "devDependencies": {
+    "jest": "^22.4.3",
+    "regenerator-runtime": "^0.11.1",
+    "sinon": "^4.4.9"
+  }
+}
