+++ packages/decoupled-logger/src/lib/any-to-string.ts
