+++ packages/decoupled-logger/src/logger-interface.ts
