+++ packages/decoupled-logger/src/logger.ts
@@ -6,7 +6,7 @@ import {
 import Chalk from 'chalk';
 import * as Transport from 'winston-transport';
 import { LoggerInterface } from './logger-interface';
-import { anyToString } from '../lib/any-to-string';
+import { anyToString } from './lib/any-to-string';
 
 export class Logger implements LoggerInterface {
