+++ packages/decoupled-logger/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "./tsconfig.base.json",
+  "exclude": [
+    "node_modules",
+    "**/*.test.ts"
+  ]
+}
