+++ packages/decoupled-logger/typings.d.ts
@@ -1,0 +1,12 @@
+declare module "*.json" {
+    const value: any;
+    export default value;
+}
+
+declare type AnyObject = {
+    [key: string]: any;
+}
+
+declare interface GenericFunction<A, R> {
+    (arg: A): R;
+}
