+++ packages/decoupled/src/generator/generator.ts
@@ -4,7 +4,7 @@
 
 import Chalk from 'chalk';
 import MetalSmith from 'metalsmith';
-import { logger } from '../logger';
+import { logger } from 'decoupled-logger';
 import path from 'path';
 
 import renderPlugin from './plugins/render-plugin';
