--- packages/decoupled/src/logger/default-logger.ts
@@ -1,3 +1,0 @@
-import { initLogger } from './init-logger';
-
-export const logger = initLogger('default');
