--- packages/decoupled/src/logger/index.ts
@@ -1,4 +1,0 @@
-export * from './init-logger';
-export * from './logger';
-export * from './default-logger';
-
