+++ packages/decoupled/src/server/server.ts
@@ -8,7 +8,7 @@ import express from 'express';
 import vhost from 'vhost';
 
 import { appPath } from '../lib';
-import { logger } from '../logger';
+import { logger } from 'decoupled-logger';
 
 import { Site } from '../site/site';
 import { getSiteIDs, getFromDecoupledJson } from '../config';
